Removed post IDs into comment admin view and added titles. Count function added to check before deleting/editing post, category, comment

+++ app/Controller/Admin/CategoriesController.php
@@ -41,6 +41,12 @@ class CategoriesController extends AppController
     {
         $title = filter_input(INPUT_POST, 'title', FILTER_SANITIZE_SPECIAL_CHARS);
         $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_SPECIAL_CHARS);
+
+        $records = $this->Category->Count($id);
+        if ($records->total == 0) {
+            $_SESSION['noRecords'] = "Aucun enregistrement correspondant trouvé / Catégorie déjà supprimée";
+            return $this->index();
+        }
         if (!empty($title)) {
             $result = $this->Category->update($id, [
                 'title' => $title,
@@ -58,7 +64,12 @@ class CategoriesController extends AppController
     public function delete()
     {
         $id = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_SPECIAL_CHARS);
+        $records = $this->Category->count($id);
         if (!empty($id)) {
+            if ($records->total == 0) {
+                $_SESSION['noRecords'] = "Aucun enregistrement correspondant trouvé / catégorie déjà supprimée";
+                return $this->index();
+            }
             $this->Category->delete($id);
             return $this->index();
         }
